ReflectionArrayIterator: rename to IterableLike, since it handles more than arrays and doesn't necessarily use reflection BindListUtil: fold into IterableLike Steal the (excellent) test cases
